own file for DnaNetworkParams
